+++ Develappers.BillomatNet/Types/Quota.cs
@@ -10,7 +10,7 @@ namespace Develappers.BillomatNet.Types
     public class Quota
     {
         public QuotaType Entity { get; set; }
-        public int Available { get; set; }
-        public int Used { get; set; }
+        public long Available { get; set; }
+        public long Used { get; set; }
     }
 }
